Update README.md
Number Game is a simple Java-based program that generates a random number in range 1 to 100. The user can make limited guesses for the generated number. The feedback gets displayed after comparing the user's guess with generated number. The user can play multiple rounds. The code terminates with a display of the user's score.